Extract search-filter from dashboard-main.js
