+++ frontend/src/js/dashboard-main.js
@@ -100,6 +100,10 @@ import {
 import {
     openConfigModal, closeConfigModal, saveConfig,
 } from './sections/config-modal.js';
+import {
+    updateTabCounts, filterKnowledge, filterHistory,
+    jumpToProject, _openHistoryHit, _reapplySearches,
+} from './sections/search-filter.js';
 
 /* --- Tabs & Cards --- */
 function togglePriMenu(wrap) {
@@ -145,7 +149,7 @@ function sortCards() {
     items.forEach(function(c) { ct.appendChild(c); });
 }
 
-var TAB_MAP = {
+export var TAB_MAP = {
     current: ['now', 'review'],
     next: ['soon', 'later'],
     backlog: ['backlog'],
@@ -500,346 +504,6 @@ async function createNotesSubdir(readmePath, parentRelToItem) {
         if (nodeId) reloadNode(nodeId); else _reloadInPlace();
     } catch (e) {
         alert('Network error: ' + e);
-    }
-}
-
-function updateTabCounts() {
-    // Primary tabs — Projects counts active items under Projects-tab
-    // priorities; Code + Knowledge are rendered with a static count from
-    // the server and don't need a client-side refresh on priority changes.
-    var projectsCount = document.querySelectorAll('#cards .card').length;
-    var projectsTab = document.querySelector('.tabs-primary .tab[data-tab="projects"] .tab-count');
-    if (projectsTab) projectsTab.textContent = '(' + projectsCount + ')';
-    // Projects sub-tabs — these filter the `#cards` grid; count per tab.
-    document.querySelectorAll('#projects-subtabs .tab').forEach(function(t) {
-        var tab = t.getAttribute('data-subtab');
-        var allowed = TAB_MAP[tab] || [];
-        var count = [].slice.call(document.querySelectorAll('.card')).filter(function(c) {
-            return allowed.indexOf(c.getAttribute('data-priority')) !== -1;
-        }).length;
-        var span = t.querySelector('.tab-count');
-        if (span) span.textContent = '(' + count + ')';
-    });
-}
-
-/* --- Search: shared helpers used by filterKnowledge + filterHistory ---
-   Tokenise on whitespace; a card matches iff every token is a substring of
-   its textContent (case-insensitive). textContent covers title, description,
-   apps, kind, tags and whatever else the server-rendered card contains, so
-   no per-field plumbing is needed. */
-function _searchTokens(q) {
-    q = (q || '').trim().toLowerCase();
-    if (!q) return [];
-    return q.split(/\s+/);
-}
-function _cardMatches(el, tokens) {
-    if (tokens.length === 0) return true;
-    var hay = (el.textContent || '').toLowerCase();
-    for (var i = 0; i < tokens.length; i++) {
-        if (hay.indexOf(tokens[i]) === -1) return false;
-    }
-    return true;
-}
-function _setEmpty(panel, cls, text) {
-    var el = panel.querySelector('.' + cls);
-    if (text == null) {
-        if (el) el.remove();
-        return;
-    }
-    if (!el) {
-        el = document.createElement('p');
-        el.className = cls;
-        panel.appendChild(el);
-    }
-    el.textContent = text;
-}
-
-/* Last query per pane — re-applied after _reloadInPlace swaps the DOM so
-   an active filter survives a background refresh. */
-var _historySearchQ = '';
-var _knowledgeSearchQ = '';
-
-/* Build an HTML snippet showing ~`radius` chars of context around the
-   first token match in `text`, with every token occurrence wrapped in
-   <mark>. Returns '' when no token matches. Snaps the cut to the
-   nearest space on either side so words aren't sliced mid-letter. */
-function _escapeHtml(s) {
-    return s.replace(/&/g, '&amp;').replace(/</g, '&lt;').replace(/>/g, '&gt;');
-}
-function _escapeRegExp(s) { return s.replace(/[.*+?^${}()|[\]\\]/g, '\\$&'); }
-function _buildSnippet(text, tokens, radius) {
-    if (!tokens.length || !text) return '';
-    var hay = text.toLowerCase();
-    var pos = -1, hitLen = 0;
-    for (var i = 0; i < tokens.length; i++) {
-        var p = hay.indexOf(tokens[i]);
-        if (p >= 0 && (pos < 0 || p < pos)) { pos = p; hitLen = tokens[i].length; }
-    }
-    if (pos < 0) return '';
-    var start = Math.max(0, pos - radius);
-    var end = Math.min(text.length, pos + hitLen + radius);
-    // Snap to nearest word boundary (whitespace) within a small margin.
-    if (start > 0) {
-        var ws = text.lastIndexOf(' ', start);
-        if (ws >= 0 && start - ws < 20) start = ws + 1;
-    }
-    if (end < text.length) {
-        var we = text.indexOf(' ', end);
-        if (we >= 0 && we - end < 20) end = we;
-    }
-    var frag = text.substring(start, end).replace(/\s+/g, ' ').trim();
-    var html = _escapeHtml(frag);
-    var re = new RegExp('(' + tokens.map(_escapeRegExp).join('|') + ')', 'gi');
-    html = html.replace(re, '<mark>$1</mark>');
-    return (start > 0 ? '…' : '') + html + (end < text.length ? '…' : '');
-}
-/* Inject/update a .match-snippet element inside a matching card. Reuses
-   the element across keystrokes to avoid DOM churn; removes it on
-   non-match or empty tokens. */
-function _setSnippet(card, tokens) {
-    var existing = card.querySelector(':scope > .match-snippet');
-    if (tokens.length === 0) {
-        if (existing) existing.remove();
-        return;
-    }
-    // Exclude the title from the snippet source so it surfaces content
-    // the user can't already see in the card's static label.
-    var titleEl = card.querySelector(':scope > .knowledge-title');
-    var titleText = titleEl ? titleEl.textContent : '';
-    var full = card.textContent || '';
-    var body = titleText && full.indexOf(titleText) === 0
-        ? full.substring(titleText.length) : full;
-    var html = _buildSnippet(body, tokens, 60);
-    if (!html) {
-        if (existing) existing.remove();
-        return;
-    }
-    if (!existing) {
-        existing = document.createElement('div');
-        existing.className = 'match-snippet';
-        card.appendChild(existing);
-    }
-    existing.innerHTML = html;
-}
-
-/* --- Generic tree filter (knowledge + history views) ---
-   Token-AND substring match against .knowledge-card textContent. Hide
-   .knowledge-group groups that end up with zero matches; re-reveal and
-   <details>-open each matching card's ancestor groups so deep matches
-   surface without showing sibling subdirs; surface an empty-state line
-   when the whole panel is empty. Matching cards get a .match-snippet
-   showing context around the first token hit. */
-function _filterTree(panel, tokens, qTrim, emptyCls, emptyMsg) {
-    var groups = panel.querySelectorAll('.knowledge-group');
-    var cards = panel.querySelectorAll('.knowledge-card');
-    if (tokens.length === 0) {
-        groups.forEach(function(g){ g.style.display = ''; });
-        cards.forEach(function(c){
-            c.style.display = '';
-            _setSnippet(c, tokens);
-        });
-        _setEmpty(panel, emptyCls, null);
-        return;
-    }
-    groups.forEach(function(g){ g.style.display = 'none'; });
-    var totalVisible = 0;
-    cards.forEach(function(c) {
-        var match = _cardMatches(c, tokens);
-        c.style.display = match ? '' : 'none';
-        _setSnippet(c, match ? tokens : []);
-        if (!match) return;
-        totalVisible += 1;
-        var anc = c.parentElement;
-        while (anc && anc !== panel) {
-            if (anc.classList && anc.classList.contains('knowledge-group')) {
-                anc.style.display = '';
-                anc.setAttribute('open', '');
-            }
-            anc = anc.parentElement;
-        }
-    });
-    _setEmpty(panel, emptyCls, totalVisible === 0 ? emptyMsg : null);
-}
-function _resetTree(panel, emptyCls) {
-    panel.querySelectorAll('.knowledge-group').forEach(function(g){ g.style.display = ''; });
-    panel.querySelectorAll('.knowledge-card').forEach(function(c){
-        c.style.display = '';
-        _setSnippet(c, []);
-    });
-    _setEmpty(panel, emptyCls, null);
-}
-
-/* --- Knowledge search --- */
-function filterKnowledge(q) {
-    _knowledgeSearchQ = q || '';
-    _persistSearch('condash.search.knowledge', _knowledgeSearchQ);
-    var panel = document.getElementById('knowledge');
-    if (!panel) return;
-    var qTrim = (q || '').trim();
-    _filterTree(panel, _searchTokens(q), qTrim,
-        'knowledge-empty', 'No knowledge pages match "' + qTrim + '".');
-}
-
-function _persistSearch(key, value) {
-    try {
-        if (!value) sessionStorage.removeItem(key);
-        else sessionStorage.setItem(key, JSON.stringify({value: value}));
-    } catch (e) {}
-}
-
-/* --- History search ---
-   Empty query → tree view (on-disk layout grouped by month).
-   Non-empty query → debounced fetch of /search-history that indexes README
-   bodies, note/text-file contents and filenames on the server, rendered as
-   a flat results list below the toolbar. */
-var _historySearchTimer = null;
-var _historySearchAbort = null;
-function filterHistory(q) {
-    _historySearchQ = q || '';
-    _persistSearch('condash.search.history', _historySearchQ);
-    var pane = document.getElementById('history-pane');
-    var tree = document.getElementById('history');
-    var results = document.getElementById('history-results');
-    if (!pane || !tree || !results) return;
-    var qTrim = _historySearchQ.trim();
-    if (!qTrim) {
-        if (_historySearchTimer) { clearTimeout(_historySearchTimer); _historySearchTimer = null; }
-        if (_historySearchAbort) { _historySearchAbort.abort(); _historySearchAbort = null; }
-        pane.classList.remove('history-pane--query');
-        results.hidden = true;
-        results.innerHTML = '';
-        tree.hidden = false;
-        return;
-    }
-    pane.classList.add('history-pane--query');
-    tree.hidden = true;
-    results.hidden = false;
-    if (_historySearchTimer) clearTimeout(_historySearchTimer);
-    _historySearchTimer = setTimeout(function(){ _runHistorySearch(qTrim); }, 150);
-}
-
-async function _runHistorySearch(q) {
-    if (_historySearchAbort) _historySearchAbort.abort();
-    _historySearchAbort = new AbortController();
-    var results = document.getElementById('history-results');
-    if (!results) return;
-    try {
-        var res = await fetch('/search-history?q=' + encodeURIComponent(q),
-                              {signal: _historySearchAbort.signal});
-        if (!res.ok) throw new Error('HTTP ' + res.status);
-        var hits = await res.json();
-        // Discard the response if the query has changed since this fetch
-        // started — another keystroke already ran a newer fetch.
-        if (_historySearchQ.trim() !== q) return;
-        _renderHistoryResults(hits, q);
-    } catch (err) {
-        if (err && err.name === 'AbortError') return;
-        results.innerHTML = '<p class="history-empty">Search failed: ' +
-            _escapeHtml(String(err && err.message || err)) + '</p>';
-    }
-}
-
-function _renderHistoryResults(hits, q) {
-    var results = document.getElementById('history-results');
-    if (!results) return;
-    if (!hits || !hits.length) {
-        results.innerHTML = '<p class="history-empty">No projects match "' +
-            _escapeHtml(q) + '".</p>';
-        return;
-    }
-    var out = [];
-    for (var i = 0; i < hits.length; i++) {
-        out.push(_historyResultBlock(hits[i]));
-    }
-    results.innerHTML = out.join('');
-}
-
-function _historyResultBlock(row) {
-    var hitsHtml = '';
-    for (var i = 0; i < row.hits.length; i++) {
-        var h = row.hits[i];
-        var pathAttr = _escapeHtml(h.path || '');
-        var labelAttr = _escapeHtml(h.label || '');
-        // Snippet already comes HTML-escaped with <mark> wrappers from the
-        // server — inject as HTML, not text.
-        var snippetHtml = h.snippet || '';
-        hitsHtml += (
-            '<li class="history-hit" ' +
-            'data-path="' + pathAttr + '" ' +
-            'data-label="' + labelAttr + '" ' +
-            'onclick="_openHistoryHit(this)">' +
-            '<span class="hit-src hit-src-' + _escapeHtml(h.source) + '">' +
-                _escapeHtml(h.label || h.source) + '</span>' +
-            '<span class="hit-snippet">' + snippetHtml + '</span>' +
-            '</li>'
-        );
-    }
-    return (
-        '<div class="history-result" ' +
-        'data-slug="' + _escapeHtml(row.slug) + '" ' +
-        'data-status="' + _escapeHtml(row.status || '') + '" ' +
-        'data-subtab="' + _escapeHtml(row.subtab || 'current') + '">' +
-        '<div class="history-result-header">' +
-            '<span class="history-result-title">' + _escapeHtml(row.title) + '</span>' +
-            '<span class="pill">' + _escapeHtml(row.kind) + '</span>' +
-            '<span class="pill pri-' + _escapeHtml(row.status) + '">' +
-                _escapeHtml(row.status) + '</span>' +
-            '<span class="history-result-month">' + _escapeHtml(row.month) + '</span>' +
-            '<button class="history-jump" onclick="jumpToProject(this)" ' +
-                'title="Open in Projects tab" aria-label="Jump to project">' +
-                '<svg viewBox="0 0 24 24" width="15" height="15" fill="none" ' +
-                'stroke="currentColor" stroke-width="2" stroke-linecap="round" ' +
-                'stroke-linejoin="round" aria-hidden="true">' +
-                '<circle cx="12" cy="12" r="9"/>' +
-                '<circle cx="12" cy="12" r="5"/>' +
-                '<circle cx="12" cy="12" r="1.2" fill="currentColor"/>' +
-                '</svg>' +
-            '</button>' +
-        '</div>' +
-        '<ul class="history-result-hits">' + hitsHtml + '</ul>' +
-        '</div>'
-    );
-}
-
-function _openHistoryHit(el) {
-    var path = el.getAttribute('data-path');
-    var label = el.getAttribute('data-label');
-    if (path) openNotePreview(path, label || path);
-}
-
-function jumpToProject(btn) {
-    var row = btn.closest('.history-result');
-    if (!row) return;
-    var slug = row.getAttribute('data-slug');
-    var sub = row.getAttribute('data-subtab') || 'current';
-    switchTab('projects');
-    switchSubtab(sub);
-    var card = document.getElementById(slug);
-    if (!card) return;
-    card.classList.remove('collapsed');
-    card.scrollIntoView({behavior: 'smooth', block: 'center'});
-    card.classList.remove('focus-flash');
-    // Re-trigger the animation by forcing a reflow before re-adding.
-    void card.offsetWidth;
-    card.classList.add('focus-flash');
-    setTimeout(function(){ card.classList.remove('focus-flash'); }, 1800);
-}
-
-/* Re-apply any active search after the DOM is swapped or the subtab
-   changes. Safe to call when the search input isn't present (other
-   primary tab active). */
-function _reapplySearches() {
-    if (_historySearchQ) {
-        var h = document.getElementById('history-search');
-        if (h) h.value = _historySearchQ;
-        // Re-run the query against the fresh DOM. In query mode this also
-        // re-fetches /search-history so newly-added files surface.
-        filterHistory(_historySearchQ);
-    }
-    if (_knowledgeSearchQ) {
-        var k = document.getElementById('knowledge-search');
-        if (k) k.value = _knowledgeSearchQ;
-        filterKnowledge(_knowledgeSearchQ);
     }
 }
